+++ src/otf2exporter.h
@@ -62,7 +62,7 @@ private:
     void exportEntities();
     void exportEntityGroups();
     void exportEvents();
-    void exportEntityEvents(int entityid);
+    void exportEntityEvents(long entityid);
 
     QMap<QString, int> inverseStringMap;
     QMap<QString, int> * attributeMap;
